+++ src/main.rs
@@ -34,10 +34,12 @@ mod args;
 mod c_format;
 mod checker;
 mod diagnostic;
+mod dict;
 mod dir;
 mod po;
 mod rules;
 mod stats;
+mod words;
 
 use clap::Parser;